grpc: change timestamp type from i32 to u64 and define Height as protobuf type

+++ contracts/src/main/kotlin/jp/datachain/corda/ibc/types/Timestamp.kt
@@ -3,4 +3,4 @@ package jp.datachain.corda.ibc.types
 import net.corda.core.serialization.CordaSerializable
 
 @CordaSerializable
-data class Timestamp(val timestamp: Int)
+data class Timestamp(val timestamp: Long)
